Normalize JSF javaagent packages
