+++ instrumentation/jsf/jsf-mojarra-1.2/javaagent/src/mojarra12Test/java/io/opentelemetry/javaagent/instrumentation/jsf/mojarra/v1_2/Mojarra12Test.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.jsf.mojarra.v1_2;
+
+import io.opentelemetry.javaagent.instrumentation.jsf.common.javax.BaseJsfTest;
+
+class Mojarra12Test extends BaseJsfTest {
+  @Override
+  protected String getJsfVersion() {
+    return "1.2";
+  }
+}
