--- instrumentation/jsf/jsf-mojarra-1.2/javaagent/src/mojarra2Test/java/io/opentelemetry/javaagent/instrumentation/mojarra/Mojarra2Test.java
@@ -1,15 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.mojarra;
-
-import io.opentelemetry.javaagent.instrumentation.jsf.common.javax.BaseJsfTest;
-
-class Mojarra2Test extends BaseJsfTest {
-  @Override
-  protected String getJsfVersion() {
-    return "2";
-  }
-}
