--- instrumentation/jsf/jsf-mojarra-3.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/mojarra/v3_0/ActionListenerImplInstrumentation.java
@@ -1,81 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.mojarra.v3_0;
-
-import static io.opentelemetry.javaagent.instrumentation.mojarra.v3_0.MojarraSingletons.instrumenter;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-
-import io.opentelemetry.context.Context;
-import io.opentelemetry.context.Scope;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
-import io.opentelemetry.javaagent.instrumentation.jsf.common.jakarta.JsfRequest;
-import jakarta.faces.event.ActionEvent;
-import javax.annotation.Nullable;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.matcher.ElementMatcher;
-
-class ActionListenerImplInstrumentation implements TypeInstrumentation {
-
-  @Override
-  public ElementMatcher<TypeDescription> typeMatcher() {
-    return named("com.sun.faces.application.ActionListenerImpl");
-  }
-
-  @Override
-  public void transform(TypeTransformer transformer) {
-    transformer.applyAdviceToMethod(
-        named("processAction"), getClass().getName() + "$ProcessActionAdvice");
-  }
-
-  @SuppressWarnings("unused")
-  public static class ProcessActionAdvice {
-
-    public static class AdviceScope {
-      private final JsfRequest request;
-      private final Context context;
-      private final Scope scope;
-
-      private AdviceScope(JsfRequest request, Context context, Scope scope) {
-        this.request = request;
-        this.context = context;
-        this.scope = scope;
-      }
-
-      @Nullable
-      public static AdviceScope start(ActionEvent event) {
-        Context parentContext = Context.current();
-        JsfRequest request = new JsfRequest(event);
-        if (!request.shouldStartSpan() || !instrumenter().shouldStart(parentContext, request)) {
-          return null;
-        }
-        Context context = instrumenter().start(parentContext, request);
-        return new AdviceScope(request, context, context.makeCurrent());
-      }
-
-      public void end(@Nullable Throwable throwable) {
-        scope.close();
-        instrumenter().end(context, request, null, throwable);
-      }
-    }
-
-    @Nullable
-    @Advice.OnMethodEnter(suppress = Throwable.class, inline = false)
-    public static AdviceScope onEnter(@Advice.Argument(0) ActionEvent event) {
-      return AdviceScope.start(event);
-    }
-
-    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class, inline = false)
-    public static void stopSpan(
-        @Advice.Thrown @Nullable Throwable throwable,
-        @Advice.Enter @Nullable AdviceScope adviceScope) {
-      if (adviceScope != null) {
-        adviceScope.end(throwable);
-      }
-    }
-  }
-}
