--- instrumentation/jsf/jsf-mojarra-3.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/mojarra/v3_0/Mojarra3Test.java
@@ -1,10 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.mojarra.v3_0;
-
-import io.opentelemetry.javaagent.instrumentation.jsf.common.jakarta.BaseJsfTest;
-
-class Mojarra3Test extends BaseJsfTest {}
