--- instrumentation/jsf/jsf-myfaces-1.2/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/myfaces/MyFacesErrorCauseExtractor.java
@@ -1,21 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.myfaces;
-
-import io.opentelemetry.javaagent.instrumentation.jsf.common.javax.JsfErrorCauseExtractor;
-import javax.el.ELException;
-
-final class MyFacesErrorCauseExtractor extends JsfErrorCauseExtractor {
-
-  @Override
-  public Throwable extract(Throwable error) {
-    error = super.extract(error);
-    while (error.getCause() != null && error instanceof ELException) {
-      error = error.getCause();
-    }
-    return error;
-  }
-}
