--- instrumentation/jsf/jsf-myfaces-3.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/myfaces/v3_0/Myfaces3Test.java
@@ -1,10 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.myfaces.v3_0;
-
-import io.opentelemetry.javaagent.instrumentation.jsf.common.jakarta.BaseJsfTest;
-
-class Myfaces3Test extends BaseJsfTest {}
